change init order in constructors

+++ c07/ex01/iter.cpp
@@ -6,7 +6,7 @@
 /*   By: hthomas <hthomas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/30 10:01:48 by hthomas           #+#    #+#             */
-/*   Updated: 2021/04/02 09:20:52 by hthomas          ###   ########.fr       */
+/*   Updated: 2021/04/10 19:29:56 by hthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	iter(T* array, int length, void (*f)(T const &))
 {
 	if (array)
 	{
-		for (size_t i = 0; i < length; i++)
+		for (int i = 0; i < length; i++)
 		{
 			f(array[i]);
 		}
